backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
torch
torchvision
huggingface_hub
python-dotenv
pydantic